requirements.txt
asgiref==3.7.1
Django==3.2.16
django-ckeditor==6.6.1
django-cleanup==8.0.0
django-debug-toolbar==4.1.0
django-grappelli==3.0.6
django-js-asset==2.1.0
django-tinymce==3.6.1
isort==5.12.0
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sorl-thumbnail==12.9.0
sqlparse==0.4.4
tzdata==2023.3
